w3c-reports/12092014/mac-wk2/runned-tests/html/semantics/embedded-content/media-elements/track/track-element/cors/022-expected.txt

CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
Actual	Expected
{
  "event": "error",
  "requests": []
}
{
  "event": "load",
  "requests": [
    {
      "cors": "http://localhost:8800",
      "cookie": "no"
    },
    {
      "cors": "null",
      "cookie": "no"
    }
  ]
}
Test id: 2710b74d-4a2a-4df1-a505-bafbeee834e3



FAIL track CORS: Anonymous, not same-origin, with headers, redirects to same-origin, with headers assert_object_equals: property "event" expected object "[object Object]" got object "[object Object]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
check_equal@http://localhost:8800/resources/testharness.js:378:28
assert_object_equals@http://localhost:8800/resources/testharness.js:390:21
checkData@http://localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/common.js?pipe=sub:128:29
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
